requirements.txt
# ==========================================
# AI DEV AGENT - DEPENDÊNCIAS MVP
# ==========================================

# === CORE ===
python-dotenv>=1.0.0
pydantic>=2.5.3
pyyaml>=6.0.1

# === LLM PROVIDERS ===
google-generativeai>=0.3.0
httpx>=0.25.2
aiohttp>=3.9.1

# === EMBEDDINGS & VECTOR STORE ===
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.1.0
transformers>=4.35.0

# === GIT & GITHUB ===
gitpython>=3.1.40
PyGithub>=2.1.1

# === CLI & UI ===
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# === TEMPLATES ===
jinja2>=3.1.2

# === UTILITIES ===
requests>=2.31.0
asyncio>=3.4.3

# === TESTING ===
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# === CODE QUALITY ===
black>=23.12.1
ruff>=0.1.9
mypy>=1.8.0
isort>=5.13.0

# === DEVELOPMENT ===
ipython>=8.18.0
ipdb>=0.13.13

# === OPTIONAL (para GPU) ===
# faiss-gpu>=1.7.4  # Descomentar se tiver GPU compatível
